added viewmore wrapper
