+++ explorer/client/src/pages/transaction-result/TxLinks.tsx
@@ -1,10 +1,8 @@
 // Copyright (c) 2022, Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import cl from 'classnames';
-import { useState, useCallback } from 'react';
-
 import Longtext from '../../components/longtext/Longtext';
+import ViewMore from '../../components/view-more/ViewMore';
 
 import type { Category } from './TransactionResultType';
 
@@ -15,25 +13,19 @@ type Addresslist = {
     category?: string;
     links: string[];
 };
+const NUMBER_OF_ITEMS_TO_SHOW = 3;
+
 function TxLinks({ data }: { data: Addresslist }) {
-    const [viewMore, setVeiwMore] = useState(false);
-    const numberOfListItemsToShow = 3;
-    const viewAll = useCallback(() => {
-        setVeiwMore(!viewMore);
-    }, [viewMore]);
     return (
         <div className={styles.mutatedcreatedlist}>
             <h3 className={styles.label}>{data.label}</h3>
             <div className={styles.objectidlists}>
                 <ul>
-                    {data.links
-                        .slice(
-                            0,
-                            viewMore
-                                ? data.links.length
-                                : numberOfListItemsToShow
-                        )
-                        .map((objId, idx) => (
+                    <ViewMore
+                        label={data.label}
+                        limitTo={NUMBER_OF_ITEMS_TO_SHOW}
+                    >
+                        {data.links.map((objId, idx) => (
                             <li key={idx}>
                                 <Longtext
                                     text={objId}
@@ -42,26 +34,8 @@ function TxLinks({ data }: { data: Addresslist }) {
                                 />
                             </li>
                         ))}
+                    </ViewMore>
                 </ul>
-                {data.links.length > numberOfListItemsToShow && (
-                    <div className={styles.viewmore}>
-                        <button
-                            type="button"
-                            className={cl([
-                                styles.moretxbtn,
-                                viewMore && styles.viewless,
-                            ])}
-                            onClick={viewAll}
-                        >
-                            {viewMore
-                                ? 'View Less'
-                                : 'View ' +
-                                  data.links.length +
-                                  ' ' +
-                                  data.label}
-                        </button>
-                    </div>
-                )}
             </div>
         </div>
     );
